Fixing/e2e test budget plan

See merge request pt-dot-internal/pama/capex-frontend!24

+++ src/pages/budget-plans/[id]/detail.tsx
@@ -17,11 +17,12 @@ import {
   useSubmitBudgetPlanItemGroups,
 } from 'modules/budgetPlanItemGroup/hook';
 import { useDecodeToken } from 'modules/custom/useDecodeToken';
+import moment from 'moment';
 import { NextPage } from 'next';
 import Link from 'next/link';
 import { useRouter } from 'next/router';
 import React, { useState } from 'react';
-import { Button, Col, Row } from 'react-bootstrap';
+import { Badge, Button, Col, Row } from 'react-bootstrap';
 import { CellProps, Column, SortingRule } from 'react-table';
 import { toast } from 'react-toastify';
 import { getAllIds, showErrorMessage } from 'utils/helpers';
@@ -102,7 +103,7 @@ const DetailBudgetPlan: NextPage = () => {
       dataHookBudgetPlanItemGroup.data
     ) as string[];
     if (ids?.length > 0) {
-      deleteBudgetPlanItemGroups(ids);
+      if (confirm('Delete selected data?')) deleteBudgetPlanItemGroups(ids);
     }
   };
 
@@ -129,6 +130,7 @@ const DetailBudgetPlan: NextPage = () => {
     if (ids?.length > 0) {
       approvalBudgetPlanItemGroups({
         idBudgetPlanItemGroups: ids,
+        // TODO: define constant
         status: 'approve',
       });
     }
@@ -158,21 +160,36 @@ const DetailBudgetPlan: NextPage = () => {
       Header: 'Total USD',
       accessor: 'totalAmountUsd',
       minWidth: 200,
+      Cell: ({ row }: CellProps<BudgetPlanItemGroup>) => {
+        return row.values.totalAmountUsd.toLocaleString('en-EN');
+      },
     },
     {
       Header: 'Total IDR',
       accessor: 'totalAmount',
       minWidth: 200,
+      Cell: ({ row }: CellProps<BudgetPlanItemGroup>) => {
+        return row.values.totalAmount.toLocaleString('id-Id');
+      },
     },
     {
       Header: 'Status',
       accessor: 'status',
-      minWidth: 200,
+      Cell: ({ row }: CellProps<BudgetPlanItemGroup>) => {
+        return (
+          <Badge className="badge--status badge--status-blue">
+            {row.values.status}
+          </Badge>
+        );
+      },
     },
     {
       Header: 'Created At',
       accessor: 'createdAt',
       minWidth: 200,
+      Cell: ({ row }: CellProps<BudgetPlanItemGroup>) => {
+        return moment(row.values.createdAt).format('YYYY-MM-DD');
+      },
     },
     {
       Header: 'Actions',
@@ -262,17 +279,20 @@ const DetailBudgetPlan: NextPage = () => {
               data={dataHookBudgetPlanItemGroup.data}
               actions={
                 <>
-                  <LoadingButton
-                    variant="red"
-                    size="sm"
-                    className="mr-2"
-                    disabled={mutationDeleteBudgetPlanItemGroup.isLoading}
-                    onClick={handleDeleteMultipleBudgetPlanItemsGroups}
-                    isLoading={mutationDeleteBudgetPlanItemGroup.isLoading}
-                  >
-                    Delete
-                  </LoadingButton>
-                  {profile?.type === UserType.Approval ? (
+                  {profile?.type !== UserType.ApprovalBudgetPlanCapex && (
+                    <LoadingButton
+                      variant="red"
+                      size="sm"
+                      className="mr-2"
+                      disabled={mutationDeleteBudgetPlanItemGroup.isLoading}
+                      onClick={handleDeleteMultipleBudgetPlanItemsGroups}
+                      isLoading={mutationDeleteBudgetPlanItemGroup.isLoading}
+                    >
+                      Delete
+                    </LoadingButton>
+                  )}
+
+                  {profile?.type === UserType.ApprovalBudgetPlanCapex ? (
                     <LoadingButton
                       size="sm"
                       className="mr-2"
